金额格式化修复

+++ app/src/main/java/com/xmkj/md/ui/adapter/BusinessFinishAdapter.java
@@ -31,7 +31,7 @@ public class BusinessFinishAdapter extends BaseQuickAdapter<OrderBean, BaseViewH
     }
 
     private String getFormatAmount(String amountStr) {
-        return "￥ " + StringUtils.numberFormat(amountStr);
+        return "￥ " + StringUtils.numberFormat(Float.valueOf(amountStr));
     }
 
 }
